+++ client/src/components/Navbar.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import logo from '../assets/reviewmatelogo.jpg';
