+++ client/src/pages/About.jsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 function About() {
   return (
     <div>
