+++ client/src/pages/StudentRegistration.jsx
@@ -1,11 +1,12 @@
-import React from "react";
+import "./reg.css";
 import { useState } from "react";
 
-const TeacherRegistration = () => {
 
+const StudentRegistration = () => {
+  
   // State to store form data
   const [data, setData] = useState({
-    teacherid: "",
+    studentid: "",
     firstname: "",
     lastname: "",
     password: "",
@@ -19,30 +20,24 @@ const TeacherRegistration = () => {
       [id]: value,
     }));
   };
+
   const handleSubmit = async (e) => {
     e.preventDefault();
 
-    // const data = {
-    //   e.target.
-    //   username: document.getElementById("username").value,
-    //   password: document.getElementById("password").value,
-    // };
-
     if (
-      data.teacherid != null &&
+      data.studentid != null &&
       data.firstname != null &&
       data.lastname != null &&
       data.password != null
     ) {
       // Ensure none of the input fields are empty
       if (
-        data.teacherid.trim() !== "" &&
+        data.studentid.trim() !== "" &&
         data.firstname.trim() !== "" &&
         data.lastname.trim() !== "" &&
         data.password.trim() !== ""
       ) {
-        const response = await fetch("/teacherreg", {
-          // Change to "/teacherreg"
+        const response = await fetch("/studentreg", {
           method: "POST",
           headers: {
             "Content-Type": "application/json",
@@ -56,7 +51,7 @@ const TeacherRegistration = () => {
       console.error("One or more fields are null");
     }
 
-    const response = await fetch("/teacherreg", {
+    const response = await fetch("/studentreg", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
@@ -66,18 +61,17 @@ const TeacherRegistration = () => {
 
     const result = await response.json();
     console.log(result);
-  }
-
+  };
   return (
     <div className="wrapper">
-      <h2>Teacher Account</h2>
+      <h2>Student Account</h2>
       <form onSubmit={handleSubmit}>
         <div className="input-box">
-          <label htmlFor="teacherid">Teacher ID:</label>
+          <label htmlFor="studentid">Student ID:</label>
           <input
             type="number"
-            id="teacherid"
-            placeholder="Enter your Teacher ID"
+            id="studentid"
+            placeholder="Enter your Student ID"
             required
           />
         </div>
@@ -108,6 +102,10 @@ const TeacherRegistration = () => {
             required
           />
         </div>
+        <div className="policy">
+          <input type="checkbox" />
+          <h3>I accept all terms & conditions</h3>
+        </div>
         <div className="input-box button">
           <input type="submit" value="Register Now" />
         </div>
@@ -116,9 +114,14 @@ const TeacherRegistration = () => {
             Already have an account? <a href="#">Login now</a>
           </h3>
         </div>
+        <div className="text">
+          <h3>
+            Creating teacher account? <a href="/teacherreg">Teacher account</a>
+          </h3>
+        </div>
       </form>
     </div>
   );
 };
 
-export default TeacherRegistration;
+export default StudentRegistration;
