+++ client/src/pages/reg.css
